Version 3.0 compiles.
Maybe it even works too
