+++ .gitignore
@@ -1,3 +1,6 @@
 /dist/
+/dist-newstyle/
 /TAGS
 /tags
+/cabal.project.local
+.stack-work
